+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - `cxpo-order-book`: Change the display of own orders - an amount link instead of a row highlight
+- `highlight-own-exchange-orders`: Change the display of own orders - an amount link instead of a row highlight
 
 ### Fixed